Updated .ps1 scripts to only delete cache and bin if switch param is passed. In the process i also got a run with the following perf:  3.570 s | 4.0841 GB/s
Times:
        3.604 s | 4.0456 GB/s
        3.902 s | 3.7359 GB/s
        3.751 s | 3.8868 GB/s
        3.570 s | 4.0841 GB/s
        3.801 s | 3.8361 GB/s
Best : 3.570 s | 4.0841 GB/s
Mean : 3.726 s | 3.9134 GB/s
Worst: 3.902 s | 3.7359 GB/s
BRC  : 3.718 s | 3.9208 GB/s

+++ bench.ps1
@@ -5,9 +5,9 @@ Param(
     [ValidateRange(1,[int]::MaxValue)]
     [int]$RunCount = 5,
 
-    [string]$Path = "C:\CodeProjects\1BillionRowChallenge\data\NoHashtag\1_000_000_000.txt"
+    [string]$Path = "C:\CodeProjects\1BillionRowChallenge\data\NoHashtag\1_000_000_000.txt",
 
-    
+    [switch]$Clean    
 )
 
 
@@ -96,17 +96,15 @@ function Format-Throughput{
 
 [Int64]$FileSize = Get-ItemPropertyValue -Path $Path -Name Length
 
-$cacheDir = [DirectoryInfo]::new(".zig-cache")
-if($cacheDir.Exists){$cacheDir | Remove-Item -Recurse -Force}
-
-$outDir = [DirectoryInfo]::new("zig-out")
-if($outDir.Exists){$outDir | Remove-Item -Recurse -Force}
-
+if($Clean){
+    $cacheDir = [DirectoryInfo]::new(".zig-cache")
+    if($cacheDir.Exists){$cacheDir | Remove-Item -Recurse -Force}
+    $outDir = [DirectoryInfo]::new("zig-out")
+    if($outDir.Exists){$outDir | Remove-Item -Recurse -Force}
+}
 zig build -freference-trace "-Doptimize=ReleaseFast"
 
 $exeFile = [FileInfo]::new('zig-out\bin\brc.exe')
-
-
 
 $times = @()
 for($i = 0; $i -lt $RunCount; $i++){
